Fail if SSH_PUBLIC_KEY missing and ensure /root/.ssh exists

+++ devsetup/scripts/gen-edpm-compute-node.sh
@@ -27,6 +27,17 @@ SSH_PUBLIC_KEY=${SSH_PUBLIC_KEY:-"../out/edpm/ansibleee-ssh-key-id_rsa.pub"}
 MAC_ADDRESS=${MAC_ADDRESS:-"$(echo -n 52:54:00; dd bs=1 count=3 if=/dev/random 2>/dev/null | hexdump -v -e '/1 "-%02X"' | tr '-' ':')"}
 IP_ADRESS_SUFFIX=${IP_ADRESS_SUFFIX:-"$((100+${EDPM_COMPUTE_SUFFIX}))"}
 
+if [ ! -f ${SSH_PUBLIC_KEY} ]; then
+    echo "${SSH_PUBLIC_KEY} is missing. Run gen-ansibleee-ssh-key.sh"
+    exit 1
+fi
+
+if sudo test -f "/root/.ssh"; then
+    sudo mkdir /root/.ssh
+    sudo chmod 700 /root/.ssh
+    sudo chcon unconfined_u:object_r:ssh_home_t:s0 /root/.ssh
+fi
+
 cat <<EOF >../out/edpm/${EDPM_COMPUTE_NAME}.xml
 <domain type='kvm'>
   <name>${EDPM_COMPUTE_NAME}</name>
